LPS-54224 Rename
